+++ kv/org.go
@@ -155,10 +155,7 @@ func (s *Service) FindOrganization(ctx context.Context, filter influxdb.Organiza
 	}
 
 	// If name and ID are not set, then, this is an invalid usage of the API.
-	return nil, &influxdb.Error{
-		Code: influxdb.EInvalid,
-		Msg:  "no filter parameters provided",
-	}
+	return nil, influxdb.ErrInvalidOrgFilter
 }
 
 func filterOrganizationsFn(filter influxdb.OrganizationFilter) func(o *influxdb.Organization) bool {